API Refactoring - API Refactoring - Update Tests - Push Tests to GitHub Actions
