+++ .github/workflows/testing.yml
@@ -29,4 +29,4 @@ jobs:
       run: |
         cd chatbot
         python manage.py migrate
-        python manage.py test users
+        python manage.py test
