+++ chatbot/chatbot/settings.py
@@ -15,6 +15,10 @@ from pathlib import Path
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
 
+# Testing Secret Key
+SECRET_KEY = '3t6w9z$C&E)H@McQfTjWnZr4u7x!A%D*G-JaNdRgUkXp2s5v8y/B?E(H+MbPeShV'
+# Testing API Key
+API_KEY = '(H+MbQeThWmYq3t6w9z$C&F)J@NcRfUj'
 
 # Quick-start development settings - unsuitable for production
 # See https://docs.djangoproject.com/en/4.2/howto/deployment/checklist/
@@ -38,7 +42,6 @@ INSTALLED_APPS = [
     # project apps
     'users',
     'contacts',
-    'api',
 ]
 
 MIDDLEWARE = [
