+++ chatbot/chatbot/urls.py
@@ -19,5 +19,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include('api.urls'), name='api'),
+    path('api/', include('users.urls')),
+    path('api/', include('contacts.urls')),
 ]
