+++ chatbot/contacts/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path, include
+from rest_framework import routers
+from .views import ContactViewSet
+
+
+router = routers.SimpleRouter()
+router.register(r'contacts', ContactViewSet, 'contacts')
+
+urlpatterns = router.urls
