+++ chatbot/users/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path, include
+from rest_framework import routers
+from .views import UserProfileViewSet
+
+
+router = routers.SimpleRouter()
+router.register(r'users', UserProfileViewSet, 'users')
+
+urlpatterns = router.urls
